param control

+++ bin/wls_top.sh
@@ -22,17 +22,28 @@ function wls_top() {
     wls_server=$1; shift 
     opt=$1
 
+    if [ -z "$wls_server" ]; then
+        echo "No server name, no run. Sorry."
+
+        echo "Usage: wls_top.sh server_name"
+        return 1
+    fi
+
     mkdir -p /tmp/$$
     tmp=/tmp/$$
 
     export wls_server
     os_pid=$(ps aux | grep java | perl -ne 'BEGIN{$wls_server=$ENV{'wls_server'};} m{\w+\s+(\d+).+java -server.+-Dweblogic.Name=$wls_server} && print "$1 "')
-
+    if [ -z "$os_pid" ]; then
+        echo "No such server."
+        return 1
+    fi
+    
     java_bin=$(dirname $(ps -o command ax -q $os_pid | grep -v 'COMMAND' | cut -f1 -d' '))
     $java_bin/jstack $os_pid | t0 | grep -z RUNNABLE | f0 >$tmp/tdump
 
     rm -f $tmp/modules
-    
+
     _threadGetOwner 'java.lang.Thread.run' >> $tmp/modules
     _threadGetOwner 'weblogic.kernel.ExecuteThread.execute' >> $tmp/modules
     _threadGetOwner 'weblogic.server.channels.ServerListenThread.selectFrom' >> $tmp/modules
